requirements.txt
Flask~=2.3.3
requests~=2.32.3
PyYAML~=6.0.1
cachetools~=5.5.1
packaging~=24.1